added superclass for analysissubmission
