+++ src/main/java/ca/corefacility/bioinformatics/irida/model/sequenceFile/SequencingObject.java
@@ -8,6 +8,7 @@ import ca.corefacility.bioinformatics.irida.model.remote.RemoteSynchronizable;
 import ca.corefacility.bioinformatics.irida.model.run.SequencingRun;
 import ca.corefacility.bioinformatics.irida.model.sample.QCEntry;
 import ca.corefacility.bioinformatics.irida.model.sample.SampleSequencingObjectJoin;
+import ca.corefacility.bioinformatics.irida.model.workflow.submission.AbstractAnalysisSubmission;
 import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSubmission;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.hibernate.envers.Audited;
@@ -68,7 +69,7 @@ public abstract class SequencingObject extends IridaResourceSupport implements M
 	@NotAudited
 	private Set<QCEntry> qcEntries;
 
-	@ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.DETACH, mappedBy = "inputFiles")
+	@ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.DETACH, mappedBy = "inputFiles", targetEntity = AbstractAnalysisSubmission.class)
 	private List<AnalysisSubmission> analysisSubmissions;
 
 	@Enumerated(EnumType.STRING)
